fix edge case infinite wait
A thread could be waiting on a condition
when the main timeout would happen.
When the timeout happens, producers and consumers stop.
However, if a consumer or producer is waiting on a condition,
and there is no other consumer or producer to signal to the condition,
then main would join indefinitely on the thread.